Make Http a module with http errors, SingleWorkspace and MultiWorkspace classes
