+++ lib/seam.rb
@@ -5,19 +5,19 @@ require_relative "seam/lts_version"
 require_relative "seam/default_endpoint"
 require_relative "seam/request"
 require_relative "seam/logger"
-require_relative "seam/client"
-require_relative "seam/client_multi_workspace"
+require_relative "seam/http"
+require_relative "seam/http_multi_workspace"
 require_relative "seam/base_client"
 require_relative "seam/base_resource"
 require_relative "seam/errors"
-require_relative "seam/seam_webhook"
+require_relative "seam/webhook"
 
 require_relative "seam/routes/resources/index"
 require_relative "seam/routes/clients/index"
 
 module Seam
   def self.new(**args)
-    Client.new(**args)
+    Http.new(**args)
   end
 
   def self.from_api_key(api_key, endpoint: nil, wait_for_action_attempt: false, debug: false)
