+++ spec/seam_client/personal_access_token_spec.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-RSpec.describe Seam::Client do
+RSpec.describe Seam::Http do
   let(:workspace_id) { "e4203e37-e569-4a5a-bfb7-e3e8de66161d" }
 
   describe "#from_personal_access_token" do
