+++ spec/seam_client/request_spec.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-RSpec.describe Seam::Client do
+RSpec.describe Seam::Http do
   let(:seam) { Seam.new(api_key: "seam_some_api_key") }
   let(:request_id) { "request_id_1234" }
